Завершение работы над функционалом

+++ src/main/java/ru/yandex/practicum/filmorate/service/film/FilmService.java
@@ -22,14 +22,11 @@ public interface FilmService {
     void deleteById(long id);
 
 
-
-
-    List<FilmResponseDto> getPopularFilms(int count);
+    List<FilmResponseDto> getPopularFilms(long count);
 
     List<FilmResponseDto> getCommonFilms(long userId, long friendId);
 
     List<FilmResponseDto> getDirectorFilms(long directorId, DirectorFilmsSortBy sortBy);
-
 
 
 }
